Fix a build failure due to missing project/lib folder
The build used to fail due to missing jforum3/lib (which is used in the
classpath path references in the build script). This commit fixes that
issue by using Ant fileset's erroronmissingdir attribute to be set to
false.

Note that this gets the build past this specific error but it fails
later due to a lot of missing jars (like JPA, Hibernate and Spring).